Added a Redirect source plugin class to use the alias field. Moved query() and fields() methods from BaseNodeSource to Redirect. Updated the migrations accordingly.
