+++ docroot/sites/sppa.uiowa.edu/modules/sppa_migrate/config/split/migrate_plus.migration.sppa_articles_redirects.yml
@@ -10,7 +10,7 @@ migration_tags:
 migration_group: sitenow_migrate
 label: 'Article Redirects'
 source:
-  plugin: sppa_articles
+  plugin: sitenow_alias_redirect
   node_type: news_story
   constants:
     redirect_prefix: 'internal:/node/'
